backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
sentence-transformers
scikit-learn
numpy
PyPDF2
google-generativeai
python-dotenv